backend/ml_service/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
gunicorn==22.0.0

# ML / data
pandas>=2.0,<4
numpy>=1.26,<3
scikit-learn>=1.5,<1.8
xgboost>=2.0,<3
joblib>=1.3,<2
